backend/requirements.txt
# DigiSoft Nexus — Backend Requirements (All Phases)
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
cryptography==42.0.8
django-ratelimit==4.1.0
celery==5.3.6
redis==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
hubspot-api-client==8.2.1
requests==2.32.3
httpx==0.27.0
Pillow==10.4.0
django-storages==1.14.4
boto3==1.34.144
django-filter==24.2
django-extensions==3.2.3
dj-database-url==2.2.0
whitenoise==6.7.0
gunicorn==22.0.0
python-slugify==8.0.4
sendgrid==6.11.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
django-debug-toolbar==4.4.2
factory-boy==3.3.1
Faker==26.0.0
